refactor : 전체 지하철 노선 보여주는 기능 코드 리팩토링
- views 폴더 생성 및 파일 이동
