+++ src/components/views/mapPrintManagerPage.js
@@ -1,4 +1,4 @@
-import MAP_MANAGER_PAGE_TEMPLATE from './template/mapPrintManagerTemplate.js';
+import MAP_MANAGER_PAGE_TEMPLATE from '../template/mapPrintManagerTemplate.js';
 
 export default function mapPrintManagerPage($element) {
   $element.innerHTML = MAP_MANAGER_PAGE_TEMPLATE();
